refactor: move utility and dice functions to modules
